.devcontainer/requirements.txt
pyyaml
ipython-autotime
# # jupyter development
# jupyter
# matplotlib
# seaborn
# plotly

# # utils
# pip-search==0.0.12